fix: wsl lm studio installer

+++ scripts/ai.sh
@@ -37,22 +37,25 @@ main() {
         esac
         ;;
       "LM Studio")
-        case "$DETECTED_DISTRO" in
-        "mac")
-          brew install --cask lm-studio
-          ;;
-        *)
-          APP_IMAGE_FILENAME="lm-studio.AppImage"
-            APP_NAME="LM Studio"
-            EXECUTABLE_PATH="/usr/bin/$APP_IMAGE_FILENAME"
-            sudo wget -cO "$EXECUTABLE_PATH" "https://lmstudio.ai/download/latest/linux/x64"
-            sudo chmod +x "$EXECUTABLE_PATH"
-            RAW_ICON="/usr/share/icons/hicolor/scalable/apps/lm-studio.webp"
-            sudo rm -rfv $RAW_ICON
-            sudo wget -cO "$RAW_ICON" "https://thanhtunguet.info/assets/img/lmstudio.webp"
-            DESKTOP_ENTRY_DIR="$HOME/.local/share/applications"
-            sudo rm -rfv "$DESKTOP_ENTRY_DIR/$APP_NAME.desktop"
-            cat << EOF > "$DESKTOP_ENTRY_DIR/$APP_NAME.desktop"
+        if [ "$IS_WSL" == "true" ]; then
+          winget.exe install -e --id ElementLabs.LMStudio
+        else
+          case "$DETECTED_DISTRO" in
+          "mac")
+            brew install --cask lm-studio
+            ;;
+          *)
+            APP_IMAGE_FILENAME="lm-studio.AppImage"
+              APP_NAME="LM Studio"
+              EXECUTABLE_PATH="/usr/bin/$APP_IMAGE_FILENAME"
+              sudo wget -cO "$EXECUTABLE_PATH" "https://lmstudio.ai/download/latest/linux/x64"
+              sudo chmod +x "$EXECUTABLE_PATH"
+              RAW_ICON="/usr/share/icons/hicolor/scalable/apps/lm-studio.webp"
+              sudo rm -rfv $RAW_ICON
+              sudo wget -cO "$RAW_ICON" "https://thanhtunguet.info/assets/img/lmstudio.webp"
+              DESKTOP_ENTRY_DIR="$HOME/.local/share/applications"
+              sudo rm -rfv "$DESKTOP_ENTRY_DIR/$APP_NAME.desktop"
+              cat << EOF > "$DESKTOP_ENTRY_DIR/$APP_NAME.desktop"
 [Desktop Entry]
 Name=$APP_NAME
 Comment=Local LLM Interface
@@ -63,10 +66,11 @@ Terminal=false
 Categories=Development;AI;Utility;
 StartupNotify=true
 EOF
-            chmod +x "$DESKTOP_ENTRY_DIR/$APP_NAME.desktop"
-            update-desktop-database "$DESKTOP_ENTRY_DIR"
-          ;;
-        esac
+              chmod +x "$DESKTOP_ENTRY_DIR/$APP_NAME.desktop"
+              update-desktop-database "$DESKTOP_ENTRY_DIR"
+            ;;
+          esac
+        fi
         ;;
     esac
     menu
